Add SYS_read, exec argv, upstream xv6 userland; fix lost-tick window
SYS_read gives the kernel a cooked console (the consoleintr slice of
kernel/console.c: echo, backspace editing, CR->NL, committed-line
delivery); the usys read() wrapper sleeps a tick between polls so
read() blocks from the caller's view. exec() now passes argv: the
kernel copies the vector+strings into a fresh area and hands
argc/argv through the new image's r0/r1 register words (crt0's call
leaves them untouched; zero-init regs mean argc=0 for plain starts).

Upstream user/ulib.c and user/echo.c compile and run unmodified;
user/printf.c is narrowed to ILP32 integer paths per the PORT.md
type policy. dma-sh gains 'run <img> [args]' (fork + exec + wait;
the vfork child only uses the private exec/exit stubs). The shell
bundle registers upstream echo + hello; typed at the silicon prompt,
'run echo booom from real silicon' executes upstream echo with
arguments through the kernel's own loader.

Also: Machine.TXPace models UART transmit pacing so kernel writes
take realistic multi-tick durations in tests and bundle verifies;
kernel SYS_write emits NL as CRNL.

The big find is a silicon-only kernel bug: a tick firing between
kenter's dispatch check and the injector retarget was never consumed,
so the injector never re-armed and the tick stream died (a RUNNABLE
process could starve forever — seen twice on hardware, ~0.1% per
syscall). kexit now re-checks the ENTRY-TIME dispatch address (not
proc[curr].pdispatch, which exec repoints mid-call); every fire is
consumed by one of four mutually covering checks. Two consecutive
full silicon passes (24/24) plus a multi-spawn interactive session
confirm. Upstream sh.c remains blocked on dmacc dynamic frames
(parsecmd/runcmd recursion) — the next compiler rung (prompts/017).

+++ xv6/user/printf.c
@@ -13,11 +13,11 @@ putc(int fd, char c)
 }
 
 static void
-printint(int fd, long long xx, int base, int sgn)
+printint(int fd, int xx, int base, int sgn) // DMA port: ILP32, no i64
 {
   char buf[20];
   int i, neg;
-  unsigned long long x;
+  unsigned int x;
 
   neg = 0;
   if (sgn && xx < 0) {
@@ -39,13 +39,13 @@ printint(int fd, long long xx, int base, int sgn)
 }
 
 static void
-printptr(int fd, uint64 x)
+printptr(int fd, uint32 x) // DMA port: ILP32
 {
   int i;
   putc(fd, '0');
   putc(fd, 'x');
-  for (i = 0; i < (sizeof(uint64) * 2); i++, x <<= 4)
-    putc(fd, digits[x >> (sizeof(uint64) * 8 - 4)]);
+  for (i = 0; i < (sizeof(uint32) * 2); i++, x <<= 4)
+    putc(fd, digits[x >> (sizeof(uint32) * 8 - 4)]);
 }
 
 // Print to the given fd. Only understands %d, %x, %p, %c, %s.
@@ -73,29 +73,29 @@ vprintf(int fd, const char *fmt, va_list ap)
       if (c0 == 'd') {
         printint(fd, va_arg(ap, int), 10, 1);
       } else if (c0 == 'l' && c1 == 'd') {
-        printint(fd, va_arg(ap, uint64), 10, 1);
+        printint(fd, va_arg(ap, uint32), 10, 1);
         i += 1;
       } else if (c0 == 'l' && c1 == 'l' && c2 == 'd') {
-        printint(fd, va_arg(ap, uint64), 10, 1);
+        printint(fd, va_arg(ap, uint32), 10, 1);
         i += 2;
       } else if (c0 == 'u') {
         printint(fd, va_arg(ap, uint32), 10, 0);
       } else if (c0 == 'l' && c1 == 'u') {
-        printint(fd, va_arg(ap, uint64), 10, 0);
+        printint(fd, va_arg(ap, uint32), 10, 0);
         i += 1;
       } else if (c0 == 'l' && c1 == 'l' && c2 == 'u') {
-        printint(fd, va_arg(ap, uint64), 10, 0);
+        printint(fd, va_arg(ap, uint32), 10, 0);
         i += 2;
       } else if (c0 == 'x') {
         printint(fd, va_arg(ap, uint32), 16, 0);
       } else if (c0 == 'l' && c1 == 'x') {
-        printint(fd, va_arg(ap, uint64), 16, 0);
+        printint(fd, va_arg(ap, uint32), 16, 0);
         i += 1;
       } else if (c0 == 'l' && c1 == 'l' && c2 == 'x') {
-        printint(fd, va_arg(ap, uint64), 16, 0);
+        printint(fd, va_arg(ap, uint32), 16, 0);
         i += 2;
       } else if (c0 == 'p') {
-        printptr(fd, va_arg(ap, uint64));
+        printptr(fd, va_arg(ap, uint32));
       } else if (c0 == 'c') {
         putc(fd, va_arg(ap, uint32));
       } else if (c0 == 's') {
